Create day1.java
Day 1. 

Not super elegant. Had a lot of trouble initially because I didn't realize that the words could overlap, and starting from one would change "twone" into "tw1" which doesn't work.

Was initially trying to build a long if-else tree, but had some problems with substring at the end of the line due to the exclusive end index being out of bounds. Instead of building a second method for end of lines or overwriting substring, looked for some help on the AoC subreddit. Someone suggested keeping the start and tail letters to not break other number-words, which worked perfectly.