Create day2.java
Some in-efficiencies in order to run both parts concurrently. 

There is probably a much more elegant way to break down the original file, and this requires the games to be in order for Part 1 since it uses the index + 1 to keep track of the game number.

Part 2 doesn't actually require you to split the games into rounds, since it doesn't care about the relationship with the other pulls in that round.  It would remove the need for an second nested loop, but it would still need it to run Part 1, so taking it out is just adding extra methods and loops.