Create day03.java
Not too happy with this one.

Very messy and bloated since it combines 2 ideas.

Part 1 Idea was to build a list of symbols, then build the numbers while checking if any digits are next to a symbol.

Part 2 became building a list of the gears, a list of the indexes of numbers, and then checking each gear for 2 surrounding numbers. There was an attempt to build a method to remove false gears with only 1 number, but it created problems and accounting for the possibility of a false gear was less code.

Part 1 could be changed to match Part 2 logic probably, but it builds the number indexes while also doing Part 1, so it might be more clean and cohesive logically but overall more code.