Create day05.java
Some pretty ugly parsing lines, but saves a lot of space. Assumes the format of the input file, but could be build with more while loops to ignore non number lines. Still requires knowledge of the order and format of the conversion tables.

Part 1 was easy when parsed.

Part 2 overloaded the java heap when brute force was attempted, but worked on the smaller example data. Hints on the subreddit suggested running it backwards; it was simple to reverse the way that I had set up Part 1. The input itself subtly suggests that by listing it as [Destination, Source, Range] I noticed.