Create day08.java
Another easy Part 1 that requires a different approach to Part 2.

Part 2 solution coded again from an idea on the AoC subreddit. Initially I had an expanded version of Part 1, but it would take too long to compute on my PC. Found out that it worked with LCM, so implemented it here. Just did a brute force LCM based on the lowest loop length. Also ended up taking a bit longer than I thought, though nowhere near as long as the full brute force. Could clean up some variables instead of reusing them for clarity of code.

The LCM approach assumes that the loop is consistent after the first Z node, which is not explicitly specified in the prompt.