Update and rename day07part1.java to day07.java
Added Part 2. Logic is controlled by part2 boolean flag, currently "off".

Still all the problems and heavy space usage of the previous version. Adapted to Part 2 using a giant if tree. There might be some kind of better math way to converting the hands, but it was faster to just hardcode all situations than figure it out. Making Jokers the weakest card was simple; moved the substitution a bit and used an if statement based on the part2 flag