Create day09.java
Part 1 was easy enough but the main 2 problems were with determining the final 0-array. At first I was using a sum based approach, but I was not accounting for negative changes, and then also discovered instances where the sum might be zero but it was not the last array. Changed to a flag based approach which works in all cases.

Part 1 was just simply inputting reversed arrays. Worked right away.