Create day10
Pretty long branches of if statements to "walk" the path. Have done some variations so some was adapted from other problems, though that introduced some errors in adapting them.

Part 1 was simple, though at first I was calculating the absolute distance of each node, not the total path distance. Was an easy fix to just add a count, and divide by two. All given input paths must have been even length since it worked.

Part 2 draws from the part one distance map, but doesn't actually need the distances-it just uses it to see if the coordinate is "on path". 

The start needs to be set for this to work, which I cheated by finding it in the file and deducing it's actual shape. It could be done by code, by finding the start and end symbols/directions and deciding based on it but it's just another if or switch statement.

There was some weirdness with the code, where the initial start if statement had some typos and therefore wasn't properly checking all sides when starting. This would only affect 3/4 of inputs since the actual "walking" loop was correct. I'm not 100% sure it's fixed and walking correctly, but it worked for the inputs that I had and tested.