Update and rename day15p1.java to day15.java
Part 2 added. Not super happy with the Array of ArrayLists, which doesn't quite feel right. Works well though.

Part 1 simply hashes the code.
Part 2 maintains the boxes as an array, and the individual boxes an an arrayList. The lenses are stored in a separate hashMap, which is updated with the lens power as needed.